==> 00_generate_lists_rw.ipynb <==
Author: KMonzella <br>
Date: 2021/10/24 <br>
Subject: Generate YouTube links<br>
Overview: In order to scrape YouTube data, we first need a list of working YouTube links. This program identifies viable youtube links using selenium. This program opens Youtube, inputs a randomly generated search topic, scrolls down the YouTube page, and pulls all the URLS it encountered. These URLS are then output as 1 CSV per search topic and combined in later programs.

# import packages
import pandas as pd 
import time
from selenium                      import webdriver 
from selenium.webdriver.common.by  import By 
from selenium.webdriver.support.ui import WebDriverWait 
from selenium.webdriver.support    import expected_conditions as EC
from random_word                   import RandomWords
from webdriver_manager.chrome      import ChromeDriverManager


# define function to open youtube, scroll down to the bottom of the page, and then pull links
def scrapeout(q):
    # q is the link for the query
    driver.get(q)

    # scroll down page
    ht=driver.execute_script("return document.documentElement.scrollHeight;")
    while True:
        prev_ht=driver.execute_script("return document.documentElement.scrollHeight;")
        driver.execute_script("window.scrollTo(0, document.documentElement.scrollHeight);")
        time.sleep(2)
        ht=driver.execute_script("return document.documentElement.scrollHeight;")
        if prev_ht==ht:
            break
        
    # once all the links are loaded, pull them and append to existing list
    time.sleep(10)
    hrefs = [video.get_attribute('href') for video in driver.find_elements_by_id("thumbnail")]

    # check total length
    print(len(hrefs))
    # check deduplicated length
    print(len(set(hrefs)))
        
    # make a dataframe
    dict = {'URL': hrefs}
    df = pd.DataFrame(dict) 
    
    # outputting the data
    fname = lambda : "youtube_urls_{}.csv".format(time.strftime("%Y%m%d-%H.%M.%S"))
    # saving to csv
    df.to_csv(fname()) 


# import chrome driver. Note that you have to download a Chrome driver if you don't already have one and save it to users/bin
driver = webdriver.Chrome(ChromeDriverManager().install())

# initialize counter
i=0

# random word function
r = RandomWords()

# initialize youtube link stem
t1 = "https://www.youtube.com/results?search_query="

while i < 50:
    print("Run number: " + str(i))
    i=i + 1
    t2 = r.get_random_word()
    t = t1.strip() + t2.strip()
    
    scrapeout(t)
