==> 01_combine_csvs.ipynb <==
Author: KMonzella <br>
Date: 2021/10/26 <br>
Subject: Combine CSVs <br>
Overview: In our 00 program, we output many CSV files of YouTube links, one from each search topic from the random word generator. In this program, we combine the output CSV files into a single CSV file of links for use.

import os
import time
import glob
import pandas as pd
os.chdir("/Users/kelleymonzella1/Desktop/Graduate_school_work/Big_Data/Final_project/Data/to_combine")

extension = 'csv'
all_filenames = [i for i in glob.glob('*.{}'.format(extension))]
print(all_filenames)


# combine all files in the list
combined_csv = pd.concat([pd.read_csv(f) for f in all_filenames ])

# keep just the URL column, not the index
lim_csv = combined_csv['URL']
lim_csv.reset_index(drop=True)

print(lim_csv.shape[0])

# drop any duplicates that might have been pulled during processing
lim_csv = lim_csv.drop_duplicates()
print(lim_csv.shape[0])

# export to csv
fname = lambda : "youtube_urls_combined_{}.csv".format(time.strftime("%Y%m%d"))
lim_csv.to_csv(fname(), index=False, encoding='utf-8-sig')
