==> 03_data_cleaning_spark.ipynb <==
Author: KMonzella <br>
Date: 2021/11/06 <br>
Subject: Data Cleaning in Spark <br>
Overview: In our 02 program, we pulled our YouTube data from links generated in earlier programs. We are now going to perform preliminary data cleaning on the raw files and output clean analysis files with all constructs except some NLP-based constructs which will be generated separately.
    

## Step 1. Program setup

import pandas                as pd
import numpy                 as np
import matplotlib.pyplot     as plt
import seaborn               as sns
from   scipy                 import stats
from pyspark.sql.functions   import *
import pyspark.sql.functions as f
from pyspark.sql.types       import IntegerType,BooleanType,DateType
from pyspark.sql.functions   import rank, col,approxCountDistinct, countDistinct

#change configuration settings on Spark 
spark = SparkSession.builder.appName('SparkBasics').getOrCreate()

conf = spark.sparkContext._conf.setAll([('spark.executor.memory', '200g'), ('spark.app.name', 'Spark Updated Conf'), ('spark.executor.cores', '8'), ('spark.cores.max', '8'), ('spark.driver.memory','30g')])

#print spark configuration settings
sc   = spark.sparkContext
spark

## Step 2: Read in CSV input file

%time df = spark.read.format("org.apache.spark.csv").option("multiline", "true").csv("/user/kmonzella/data/youtube_raw_20211120.csv", inferSchema=True, header=True, quote='\"', sep = ",", escape='"')
df.printSchema() 


# check what data look like
df.show(3)

# total record count
df.count()

## Step 3: Preliminary cleaning

### A. Check for duplicates

df = df.dropDuplicates(['watch_url'])

df.count()

### B. Check for other missing data

# check missingness
df.select([count(when(isnan(c) | col(c).isNull(), c)).alias(c) for c in df.columns]).show()

# there are the overflow from the cut off transcripts. As long as this is low prevalence, drop
percent_nonmissing = df.filter(df.watch_url.startswith('https://youtube.com/watch?v=')).count() * 100 / df.count()
print(percent_nonmissing)

df1 = df.filter(df.watch_url.startswith('https://youtube.com/watch?v='))


# recheck missingness
df1.select([count(when(isnan(c) | col(c).isNull(), c)).alias(c) for c in df1.columns]).show()

# address other missingness - transcripts

df1 = df1.withColumn("transcript", f.when(col("transcript")=='No transcript available', None).otherwise(col("transcript")))


df1.select("transcript").show()

### C. Check and change data types

# --- numeric

cols = ["length", "rating", "views"]
for col in cols:
    df1 = df1.withColumn(
            col,
            f.col(col).cast("double"))


# --- dates

df1 = df1.withColumn("publish_date_new", to_date("publish_date", "yyyyMMdd"))

print(df1.select(["publish_date", "publish_date_new"]).show(10)) 

df1 = df1.drop("publish_date")

# --- booleans

df1 = df1.withColumn("age_restricted",  df1.age_restricted.cast(BooleanType()))

## Step 4: Create new variables

### A. Indicators of whether there is a song in the video

df1 = df1.withColumn("has_music", f.when(f.col("metadata").contains('Song'), 1.0).otherwise(0.0))

df1.filter("has_music==1").select(["has_music", "metadata"]).show(10)
df1.filter("has_music==0").select(["has_music", "metadata"]).show(10)

# transcript has sound effects
df1 = df1.withColumn("has_sounds", f.when(f.col("transcript").contains('[Music]'), 1.0).otherwise(0.0))
df1.filter("has_sounds==1").select(["has_sounds", "transcript"]).show(10)
df1.filter("has_sounds==0").select(["has_sounds", "transcript"]).show(10)

### B. Indicators related to streaming quality

# not needed at this time

### C. Indicators related to duration - binary

df1 = df1.withColumn("duration_long",   f.when(f.col('length') >= 1800, 1).otherwise(0))
df1 = df1.withColumn("duration_med",    f.when((f.col('length') < 1800) & (f.col('length')>=600), 1).otherwise(0))
df1 = df1.withColumn("duration_short",  f.when(f.col('length') < 600,   1).otherwise(0))

df1.filter("duration_long==1").select(["length", "duration_long", "duration_med", "duration_short"]).show()
df1.filter("duration_long==0").select(["length", "duration_long", "duration_med", "duration_short"]).show()

### E. Keywords

# count number of keywords - perhaps more search terms helps you get more views
df1 = df1.withColumn('n_keywords', f.size(f.split(f.col("keywords"), r"\|")) + 1)

df1.select(["keywords", "n_keywords"]).show()

### G. Dates

# --- post covid 

# create a covid inicator. Assume official start of COVID-19 pandemic as March 27, 2020 
# this is the date the CARES act was implemented
df1 = df1.withColumn('post_covid', f.when(f.col('publish_date_new') > "2020-03-26", 1).otherwise(0))
df1.select(["post_covid", "publish_date_new"]).show()

### H. Author

from pyspark.sql import Window
w = Window.partitionBy('author')
df1 = df1.withColumn('n_posts_by_author', f.count('author').over(w))
df1.agg(f.min(f.col("n_posts_by_author")),  f.max(f.col("n_posts_by_author"))).show()

w = Window.partitionBy('channel')
df1 = df1.withColumn('n_posts_by_channel', f.count('channel').over(w))
df1.agg(f.min(f.col("n_posts_by_channel")),  f.max(f.col("n_posts_by_channel"))).show()

df1.sort("author").select(["author", "n_posts_by_author"]).show()

## Step 5: Transformations to outcome variable

df1 = df1.withColumn('views_high', f.when(f.col('views') > 10000, 1).otherwise(0))
df1.select(["views_high", "views"]).show()

# categorical views - high medium low
df1 = df1.withColumn("views_cat", f.when(f.col('views') > 10000, 3).otherwise(
     f.when(f.col("views") > 1000, 2).otherwise(1)))
    
df1.filter("views_cat==3").select(["views_cat", "views"]).show(10)
df1.filter("views_cat==2").select(["views_cat", "views"]).show(10)
df1.filter("views_cat==1").select(["views_cat", "views"]).show(10)

# check min, max, median, and average values for references
df1.agg(f.min(f.col("views")), f.max(f.col("views")), f.avg(f.col("views"))).show()

## Step 6: Output df 

df1.printSchema()

%time df1.coalesce(1).write.mode('overwrite').option('header','true').csv('/user/kmonzella/data/youtube_clean.csv')

#hdfs dfs -copyToLocal /user/kmonzella/data/youtube_clean.csv ~/data/
