==> 05_big_data_platforms_nlp_trending_data_part_1.ipynb <==
!wget http://setup.johnsnowlabs.com/colab.sh -O - | bash
!pip install transformers
!pip install spacy
!pip install yake

from google.colab import drive
drive.mount('/content/drive')

import os

#!unzip '/content/drive/MyDrive/all_the_news.zip'

from glob import glob
fp = [f for f in glob('/content/drive/**', recursive=True) if 'youtube_clean_' in f and '.json' in f][0]
fp_news = '/content/drive/MyDrive/all-the-news-2-1.csv'

import sparknlp
from pyspark.sql import SparkSession

spark = sparknlp.start()
spark = SparkSession.builder \
    .appName("Spark NLP")\
    .master("local[4]")\
    .config("spark.driver.memory","16G")\
    .config("spark.driver.maxResultSize", "0") \
    .config("spark.kryoserializer.buffer.max", "2000M")\
    .config("spark.jars.packages", "com.johnsnowlabs.nlp:spark-nlp_2.12:3.3.2")\
    .getOrCreate()

from pyspark.sql.functions import col, year, month, udf, concat, explode
from pyspark.sql.types import StringType, ArrayType

preprocess_article = udf(lambda x: pre_process(x), ArrayType(StringType()) )

news = spark.read.format("org.apache.spark.csv").option("multiline", "true").csv(fp_news, inferSchema=True, header=True, quote='"', sep = ",", escape='\"')
news.show(3)

from sparknlp.annotator import *
from sparknlp.base import *

document_assembler = DocumentAssembler().setInputCol('article').setOutputCol('document')

sentence_detector = SentenceDetector() \
    .setInputCols(['document']) \
    .setOutputCol('sentences')

tokenizer = Tokenizer() \
    .setInputCols(['sentences']) \
    .setOutputCol('tokens') \
    .setContextChars(['(', ')', '?','!','.',','])

keywords = YakeKeywordExtraction() \
    .setInputCols('tokens') \
    .setOutputCol('keywords') \
    .setMinNGrams(2) \
    .setMaxNGrams(5) \
    .setNKeywords(5) \
    .setStopWords(StopWordsCleaner().getStopWords())


pipeline_news = Pipeline(stages=[
    document_assembler,
    sentence_detector,
    tokenizer,
    keywords,
])
model_pipeline = pipeline_news.fit(news)
news_transformed = model_pipeline.transform(news)

extract_keywords = udf(lambda x: list(set([j.result for j in x])), ArrayType(StringType()))
news_transformed = news_transformed.withColumn('keywords_only', extract_keywords(news_transformed.keywords ))
news_transformed = news_transformed.withColumn('month', news_transformed.month.cast('integer'))
news_transformed = news_transformed.withColumn('year_month', concat(news_transformed.year, news_transformed.month ))
news_transformed_exploded = news_transformed.select(news_transformed.year_month, explode(news_transformed.keywords_only))
news_transformed_keyword_count = news_transformed_exploded.groupBy(['year_month','col']).count()

news_transformed_keyword_count.show(3)

news_transformed_keyword_count.coalesce(1).write.format('json').save('/content/drive/MyDrive/news_keywords_count.json')
