==> 07_big_data_platforms_part3.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

!wget http://setup.johnsnowlabs.com/colab.sh -O - | bash

import pickle
input_df_train, input_df_test, train_y, test_y = pickle.load(open('/content/drive/MyDrive/bdp_project_final_train_test.p', mode='rb'))

for col in ['rating', 'duration_long','n_posts_by_author']:
    input_df_train.drop(col, axis=1, inplace=True)
    input_df_test.drop(col, axis=1, inplace=True)

input_df_train.fillna(False, inplace=True)
input_df_test.fillna(False, inplace=True)

def normalize(X):
    for col in ['n_keywords', 'n_posts_by_channel']:
        X[col] = (X[col]-X[col].mean() )/X[col].std()

    return X

from sklearn.metrics import classification_report, accuracy_score, confusion_matrix, f1_score, precision_score, recall_score

def score(test_y, ypred_test_final):
    print(confusion_matrix(test_y, ypred_test_final))
    print(classification_report(test_y, ypred_test_final))
    print('accuracy_score:  '+str(accuracy_score(test_y, ypred_test_final)))
    print('f1_score:        '+str(f1_score(test_y, ypred_test_final)))
    print('precision_score: '+str(precision_score(test_y, ypred_test_final)))
    print('recall_score:    '+str(recall_score(test_y, ypred_test_final)))

import pandas as pd
from sklearn.linear_model import LogisticRegression

lmod = LogisticRegression()
lmod.fit(normalize(input_df_train), train_y)
ypred_test_final = lmod.predict(normalize(input_df_test))

score(test_y, ypred_test_final)
coefs = pd.DataFrame(lmod.coef_).T
coefs.index = input_df_train.columns
coefs.columns=['coef']
coefs.sort_values('coef')

from sklearn.ensemble import RandomForestClassifier
rf = RandomForestClassifier()
rf.fit(input_df_train, train_y)
ypred_test_final = rf.predict(input_df_test)


score(test_y, ypred_test_final)
feature_importance = pd.DataFrame(rf.feature_importances_)
feature_importance.index=input_df_train.columns
feature_importance.columns=['feature_importance']
feature_importance.sort_values('feature_importance', ascending=False)

from sklearn.svm import SVC
svc = SVC(kernel='linear')
svc.fit(normalize(input_df_train), train_y)
ypred_test_final = svc.predict(normalize(input_df_test))

score(test_y, ypred_test_final)
svc_coefs = pd.DataFrame(svc.coef_).T
svc_coefs.index = input_df_train.columns
svc_coefs.columns=['coef']
svc_coefs.sort_values('coef',ascending=False)
